# proyecto3/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Banco XYZ Services
  description: API para gestionar transacciones del Banco XYZ
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
...
  /cuentas:
    get:
      tags:
        - cuenta
      summary: Listar cuentas
      operationId: listAccounts.
      responses:
        '200':
          description: Lista de cuentas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
        default:
          description: Error en la obtención de cuentas
    post:
      tags:
        - cuenta
      summary: Crear una cuenta
      operationId: registerAccount
      requestBody:
        description: Registrar una cuenta
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRequest"
      responses:
        '200':
          description: Cuenta registrada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        default:
          description: Error en la creacion de cuenta
  /cuentas/{id}:
    get:
      tags:
        - cuenta
      summary: Obtener datos de una cuenta
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cuenta obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        default:
          description: Error en obtener datos de la cuenta
    delete:
      tags:
        - cuenta
      summary: Eliminar cuenta por su ID
      operationId: deleteAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Cuenta eliminada
        default:
          description: Error en eliminar la cuenta
  /cuentas/{numeroCuenta}/depositar:
    put:
      tags:
        - cuenta
      summary: Realizar un depósito en una cuenta bancaria
      operationId: depositAccount
      parameters:
        - name: numeroCuenta
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Monto a depositar
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Amount"
      responses:
        '200':
          description: Depósito realizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        default:
          description: Error en el depósito en la cuenta
  /cuentas/{numeroCuenta}/retirar:
    put:
      tags:
        - cuenta
      summary: Realizar un retiro de una cuenta bancaria
      operationId: withdrawAccount
      parameters:
        - name: numeroCuenta
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Monto a depositar
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Amount"
      responses:
        '200':
          description: Retiro realizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        default:
          description: Error en el retiro en la cuenta
...
components:
  schemas:
...
    AccountRequest:
      type: object
      properties:
        id:
          type: string
          description: ID único de la cuenta
        numeroCuenta:
          type: string
          description: Numero de la cuenta
        saldo:
          type: number
          format: double
          description: Saldo de la cuenta
        tipoCuenta:
          type: string
          description: Tipo de la cuenta (AHORROS, CORRIENTE)
        clienteId:
          type: string
          description: ID del cliente de la cuenta
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          description: ID único de la cuenta
        numeroCuenta:
          type: string
          description: Numero de la cuenta
        saldo:
          type: number
          format: double
          description: Saldo de la cuenta
        tipoCuenta:
          type: string
          description: Tipo de la cuenta (AHORROS, CORRIENTE)
        clienteId:
          type: string
          description: ID del cliente de la cuenta
    Amount:
      type: object
      properties:
        monto:
          type: number
          format: double
          description: Monto a depositar o retirar
          example: 100.0
...